revertccl: move RevertSpansFanout to revertccl
Release note: none.
Epic: none.
