+++ pkg/BUILD.bazel
@@ -72,6 +72,7 @@ ALL_TESTS = [
     "//pkg/ccl/partitionccl:partitionccl_test",
     "//pkg/ccl/pgcryptoccl/pgcryptocipherccl:pgcryptocipherccl_test",
     "//pkg/ccl/pgcryptoccl:pgcryptoccl_test",
+    "//pkg/ccl/revertccl:revertccl_test",
     "//pkg/ccl/schemachangerccl:schemachangerccl_test",
     "//pkg/ccl/serverccl/adminccl:adminccl_test",
     "//pkg/ccl/serverccl/diagnosticsccl:diagnosticsccl_test",
@@ -899,6 +900,8 @@ GO_TARGETS = [
     "//pkg/ccl/pgcryptoccl:pgcryptoccl",
     "//pkg/ccl/pgcryptoccl:pgcryptoccl_test",
     "//pkg/ccl/plpgsqlccl:plpgsqlccl",
+    "//pkg/ccl/revertccl:revertccl",
+    "//pkg/ccl/revertccl:revertccl_test",
     "//pkg/ccl/schemachangerccl:schemachangerccl",
     "//pkg/ccl/schemachangerccl:schemachangerccl_test",
     "//pkg/ccl/securityccl/fipsccl:fipsccl",
